requirements.txt
requests
pandas
python-dotenv
zappix
pyzabbix
black
